requirements.txt
fastapi==0.115.11
langchain==0.3.21
langchain_community==0.3.20
langchain_core==0.3.47
langchain_openai==0.3.9
langgraph==0.3.18
pydantic==2.10.6
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.43.2
typing_extensions==4.12.2
uvicorn==0.34.0
